changed structure, once we refactor individual pages, these folders may populate
