+++ demo-drone-app/src/components/mapThumbnail.vue
@@ -32,7 +32,7 @@
 	import Vue from 'vue';
 	import Vuetify from 'vuetify'
 	import router from '@/router'
-	import API from '../../mixins/API.js'
+	import API from '@/mixins/API.js'
 
 	Vue.use(VueGoogleMaps, {
     load: {
