+++ demo-drone-app/src/router/index.js
@@ -2,21 +2,21 @@ import Vue from 'vue'
 import Router from 'vue-router'
 
 const routerOptions = [
-  { path: '/', component: 'LandingPage' },
-  { path: '/login', component: 'Login'},
-  { path: '/map', component: 'MapPage' },
-  { path: '/newflight', component: 'NewFlightsPage' },
-  { path: '/settings', component: 'Settings'},
-  { path: '/homepage', component: 'HomePage'},
-  { path: '/drones', component: 'Drones'},
-  { path: '/flights', component: 'FlightsPage'},
-  { path: '/education', component: 'EducationPage'}
+  { path: '/', view: 'LandingPage' },
+  { path: '/login', view: 'LoginPage'},
+  { path: '/map', view: 'FlightDetailsPage' },
+  { path: '/newflight', view: 'NewFlightPage' },
+  { path: '/settings', view: 'SettingsPage'},
+  { path: '/homepage', view: 'HomePage'},
+  { path: '/drones', view: 'DronesPage'},
+  { path: '/flights', view: 'FlightsPage'},
+  { path: '/education', view: 'EducationPage'}
 ]
 
 const routes = routerOptions.map(route => {
   return {
     path: route.path,
-    component: () => import(`@/components/${route.component}.vue`)
+    component: () => import(`@/views/${route.view}/${route.view}.vue`)
   }
 })
