+++ demo-drone-app/src/views/EducationPage/EducationPage.vue
