+++ demo-drone-app/src/views/NewFlightPage/NewFlightPage.vue
@@ -323,7 +323,7 @@
   import Vue from 'vue';
   import axios from 'axios'
   import VueAxios from 'vue-axios'
-  import API from '../mixins/API.js'
+  import API from '@/mixins/API.js'
   import moment from 'moment'
   import router from '@/router'
