+++ demo-drone-app/src/views/SettingsPage/settingsLicenses.vue
